+++ src/stats/Timer.cpp
@@ -35,7 +35,7 @@ void Timer::stop()
     (end - begin);
 
   no++;
-  sum += delta.count();
+  sum += static_cast<double>(delta.count());
 }
